// src/core/RecyclerListView.d.ts
export interface OnRecreateParams {
    lastOffset?: number;
}
export interface RecyclerListViewProps {
    layoutProvider: BaseLayoutProvider;
    dataProvider: DataProvider;
    rowRenderer: (type: string | number, data: any, index: number) => JSX.Element | JSX.Element[] | null;
    contextProvider?: ContextProvider;
    renderAheadOffset?: number;
    isHorizontal?: boolean;
    onScroll?: (rawEvent: ScrollEvent, offsetX: number, offsetY: number) => void;
    onRecreate?: (params: OnRecreateParams) => void;
    onEndReached?: () => void;
    onEndReachedThreshold?: number;
    onVisibleIndexesChanged?: TOnItemStatusChanged;
    renderFooter?: () => JSX.Element | JSX.Element[] | null;
    externalScrollView?: {
        new (props: ScrollViewDefaultProps): BaseScrollView;
    };
    initialOffset?: number;
    initialRenderIndex?: number;
    scrollThrottle?: number;
    canChangeSize?: boolean;
    distanceFromWindow?: number;
    useWindowScroll?: boolean;
    disableRecycling?: boolean;
    forceNonDeterministicRendering?: boolean;
    extendedState?: object;
    itemAnimator?: ItemAnimator;
    optimizeForInsertDeleteAnimations?: boolean;
    style?: object;
    scrollViewProps?: object;
}
export interface RecyclerListViewState {
    renderStack: RenderStack;
}
export default class RecyclerListView<P extends RecyclerListViewProps, S extends RecyclerListViewState> extends React.Component<P, S> {
    static defaultProps: {
        canChangeSize: boolean;
        disableRecycling: boolean;
        initialOffset: number;
        initialRenderIndex: number;
        isHorizontal: boolean;
        onEndReachedThreshold: number;
        distanceFromWindow: number;
        renderAheadOffset: number;
    };
    static propTypes: {};
    private _onEndReachedCalled;
    protected _virtualRenderer: VirtualRenderer;
    private _initComplete;
    private _relayoutReqIndex;
    private _params;
    private _layout;
    private _pendingScrollToOffset;
    private _tempDim;
    private _initialOffset;
    private _cachedLayouts?;
    private _scrollComponent;
    private _defaultItemAnimator;
    constructor(props: P);
    componentWillReceiveProps(newProps: RecyclerListViewProps): void;
    componentDidUpdate(): void;
    componentWillUnmount(): void;
    componentWillMount(): void;
    scrollToIndex(index: number, animate?: boolean): void;
    scrollToItem(data: any, animate?: boolean): void;
    getLayout(index: number): Layout | undefined;
    scrollToTop(animate?: boolean): void;
    scrollToEnd(animate?: boolean): void;
    scrollToOffset(x: number, y: number, animate?: boolean): void;
    updateRenderAheadOffset(renderAheadOffset: number): boolean;
    getRenderAheadOffset(): number;
    getCurrentScrollOffset(): number;
    findApproxFirstVisibleIndex(): number;
    render(): JSX.Element;
    private _checkAndChangeLayouts(newProps, forceFullRender?);
    private _refreshViewability();
    private _queueStateRefresh();
    private _onSizeChanged(layout);
    private _renderStackWhenReady(stack);
    private _initTrackers();
    private _assertDependencyPresence(props);
    private _assertType(type);
    private _dataHasChanged(row1, row2);
    private _renderRowUsingMeta(itemMeta);
    private _onViewContainerSizeChange(dim, index);
    private _checkExpectedDimensionDiscrepancy(itemRect, type, index);
    private _generateRenderStack();
    private _onScroll(offsetX, offsetY, rawEvent);
    private _processOnEndReached();
}